Stabilizing tests so it can be run on push (#4)
Part of the env variables are acting up when running on workflows, this is understandable because the environment does have the variables that otherwise will be empty and envalid will provide the testOnly to make the test work. In this PR I'm overriding the part of the env variables that destabilize the tests and it will allow these to run on push (master) to get test coverage working.