Adding debugging mode  (#11)
this PR provides a `DEBUG` option to allow the action to display verbose information.

It has been helpful when debugging issues. I thought it will be useful to leave it as a feature.

I've also updated the README for more information about what values can DEBUG take, please take a look at this package, which is used underneath.