Adding `errorLevel` to rules (#62)
Adding `errorLevel` will allow rules to fail the workflow if these are not applied. 

It will be handy to create different workflows with really specific intents. The first use case for this is the ability to fail a check when the PR does not have the right template. but it could be expanded to anything that is meant to be used for. 

* adding pull request template as well.