Add `status` action (#106)
**Issue Reference**:  fixes #71 

## Description

One good enhancement of failing a workflow is to just have rules that can fail or succeed on a  commit status.
this PR adds that feature. 

## Motivation and Context

We have found that failing the entire step when using `errorLevel: error` might be too blunt and perhaps we could a status as a way to display individual rules when `action` is set to `status`

## How Has This Been Tested?
unit test and changing validation_template to use this feature moving forward.