switching to esbuild (#244)
*Description
replacing ncc in favor of esbuild

Motivation and Context
ncc is having issues packing some modules that don't conform with esmodules. producing issues, in particular with octokit/rest.js and octokit/endpoint.js that brings is-plain-object package and universal-agent packages with break the build when updated.

This is happening mainly because ncc in some cases is pulling from the web dist instead of the module or main field in the dependencies, causing these issues.

esbuild is more configurable, and we can specifically point to the entry point we care about. (main, module)