current rule structure. with new class design (#340)
I've unified the Rules interactions with the system into 2 classes. that will serve us as a simplified API. 

`Rules` and `MatchingRules` will allow us to work on them to bring the new rules structures in place without creating confusion or changing (much) actions and the index while we rework things.