Execute status rule if every rule fails (#531)
<!--- IMPORTANT: Please do not create a Pull Request without [creating
an issue first](https://github.com/gagoar/use-herald-action/issues/new)
-->

<!--- write down the issue related to this  PR-->

**Issue
Reference**:[#530](https://github.com/gagoar/use-herald-action/issues/530)

## Description
i updated the getMatchingRules method to return all the status rules and
also the validation to update the status to be success or failure
depending on the match result and not if it exists on the Matching rules
<!--- Describe your changes in detail -->

## Motivation and Context
it fixes this [Issue
530](https://github.com/gagoar/use-herald-action/issues/530)
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
i've tested it on my private repo where i have just one status rule
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how
-->
<!--- your change affects other areas of the code, etc. -->

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.